style: Move `ApplyState` over to `node-runtime`
This type does not appear to be particularly useful outside of
`node-runtime` (`node-runtime` is its only consumer, so the only good
reason to construct it is for `node-runtime` to use) and it introduces a
dependency edge from near-primitives to near-vm-runner which we might as
well avoid.
