+++ core/primitives/src/runtime/mod.rs
@@ -1,2 +1,1 @@
-pub mod apply_state;
 pub mod migration_data;
