+++ runtime/runtime/src/state_viewer/mod.rs
@@ -1,17 +1,18 @@
-use crate::near_primitives::version::PROTOCOL_VERSION;
+use crate::actions::execute_function_call;
+use crate::ext::RuntimeExt;
 use crate::receipt_manager::ReceiptManager;
-use crate::{actions::execute_function_call, ext::RuntimeExt};
+use crate::ApplyState;
 use near_crypto::{KeyType, PublicKey};
 use near_parameters::RuntimeConfigStore;
 use near_primitives::account::{AccessKey, Account};
 use near_primitives::borsh::BorshDeserialize;
 use near_primitives::hash::CryptoHash;
 use near_primitives::receipt::ActionReceipt;
-use near_primitives::runtime::apply_state::ApplyState;
 use near_primitives::runtime::migration_data::{MigrationData, MigrationFlags};
 use near_primitives::transaction::FunctionCallAction;
 use near_primitives::trie_key::trie_key_parsers;
 use near_primitives::types::{AccountId, EpochInfoProvider, Gas};
+use near_primitives::version::PROTOCOL_VERSION;
 use near_primitives::views::{StateItem, ViewApplyState, ViewStateResult};
 use near_primitives_core::config::ViewConfig;
 use near_store::{get_access_key, get_account, get_code, TrieUpdate};
